Special on ctrl keys behavior implemented
When the user uses the left,right or backspace buttons while pressing the ctrl the caret moves to the next or previous special character as a real input would do, also it removes all the characters until the previous special character, as a real input would do it as well.

Also the caret positioning after adding a character is now managed by the append method, this due to conflicts using the ngfor directive to render the characters.