First version of editable radical roots created
The first version of the editable radical roots has been created, also the root subcomponent known as root main container has been created, so the root component works as a kind of super container just as the fraction component does.

The term argument component has been created as well, this component will serve as a container for all kind of function arguments, such as roots in this case, logarithms, defined integrals and so on.